Create requirementsgooglecollab.txt
error encountered using the requirement.txt in google cola due to some packages being pre-installed in collab. 

providing an alternative requirement txt to enable working in collar. 
"The error you're encountering is related to the attempt to install tensorflow-gpu from your requirements.txt file. The issue seems to be that the version of tensorflow-gpu listed (or the lack of a specific version) is not installing properly in the Google Colab environment. This can happen for a variety of reasons, such as compatibility issues with the environment or incorrect package specifications."